+++ frontend/src/components/Table/ListActiveProj.module.css
@@ -1,0 +1,92 @@
+.container,
+.groupContainer {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(10%, 1fr));
+  grid-auto-flow: column;
+  justify-content: start;
+  align-items: center;
+  width: 100%;
+  /* padding: 10px; */
+}
+
+.groupContainer:hover {
+  background-color: #f0f0f0;
+}
+.content {
+  min-height: 25rem;
+  max-height: 25rem;
+  overflow-x: auto; /* Make it scrollable */
+}
+
+.card {
+  /* display: flex; */
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: stretch;
+  padding: 0 0 20px 0;
+  max-width: 80rem;
+  min-height: 29rem;
+  max-height: 29rem;
+  margin: 0 auto;
+  flex-wrap: wrap;
+  width: auto; /* Set a fixed width */
+  /* overflow-x: auto; Make it scrollable */
+}
+
+.centerTextName {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  color: #9c7b16;
+  font-weight: 500;
+  cursor: pointer;
+  font-size: 0.8em;
+}
+
+.centerTextName,
+.centerTextOverallRating:hover {
+  font-weight: 700;
+}
+
+.centerText,
+.centerTextOverallRating {
+  display: flex;
+  flex-direction: column; /* Stack the items vertically */
+  justify-content: center; /* Center horizontally */
+  align-items: center; /* Center vertically */
+  height: 100%;
+  color: #292929;
+  font-size: 0.85em;
+  text-align: center; /* Center the text */
+  padding-right: 8px;
+  padding-left: 3px;
+}
+.centerTextOverallRating {
+  color: red;
+  padding-left: 30px;
+}
+.clickable {
+  cursor: pointer;
+  transition: color 0.3s ease;
+}
+
+.clickable:hover {
+  color: #9c7b16;
+}
+
+.pagination {
+  margin-top: 10px;
+  display: flex;
+  justify-content: flex-end;
+}
+
+.pagination span {
+  cursor: pointer;
+}
+
+.disabled {
+  color: #ccc; /* Change the color to gray or a color of your choice */
+  cursor: not-allowed; /* Change the cursor to indicate it's not clickable */
+  pointer-events: none; /* Make the button non-interactable */
+}
